+++ src/datasource/FarmerPlotDatasource.ts
@@ -1,12 +1,12 @@
-import { httpClient } from "../config/develop-config";
+import { httpClient, BASE_URL } from "../config/develop-config";
 import { FarmerPlotEntity } from "../entities/FarmerPlotEntities";
-const API_URL = `https://api-dev-dnds.iconkaset.com`;
+
 
 export class FarmerPlotDatasource {
   static insertFarmerPlot(data: FarmerPlotEntity): Promise<any> {
     delete data.id;
     return httpClient
-      .post(API_URL + "/farmer-plot", data)
+      .post(BASE_URL + "/farmer-plot", data)
       .then((response) => {
         return response.data;
       })
@@ -16,7 +16,7 @@ export class FarmerPlotDatasource {
   }
   static updateFarmerPlot(data: FarmerPlotEntity): Promise<any> {
     return httpClient
-      .patch(API_URL + "/farmer-plot/" + data.id, data)
+      .patch(BASE_URL + "/farmer-plot/" + data.id, data)
       .then((response) => {
         return response.data;
       })
@@ -26,7 +26,7 @@ export class FarmerPlotDatasource {
   }
   static deleteFarmerPlot(id?: string): Promise<any> {
     return httpClient
-      .delete(API_URL + "/farmer-plot/" + id)
+      .delete(BASE_URL + "/farmer-plot/" + id)
       .then((response) => {
         return response.data;
       })