+++ src/entities/DronerEntities.ts
@@ -21,6 +21,7 @@ export interface DronerEntity {
   idNo: string;
   telephoneNo: string;
   status: string;
+  reason : string[];
   expYear: number;
   expMonth: number;
   expPlant: string[];
@@ -40,6 +41,7 @@ export const DronerEntity_INIT: DronerEntity = {
   idNo: "",
   telephoneNo: "",
   status: "",
+  reason:[""],
   expYear: 0,
   expMonth: 0,
   expPlant: [""],
@@ -73,7 +75,7 @@ export const CreateDronerEntity_INIT: CreateDronerEntity = {
   lastname: "",
   idNo: "",
   telephoneNo: "",
-  status: "OPEN",
+  status: "PENDING",
   expYear: 0,
   expMonth: 0,
   expPlant: [""],