+++ src/pages/droner/DroneList.tsx
@@ -1,77 +1,83 @@
-import {
-  Avatar,
-  Badge,
-  Button,
-  Col,
-  Input,
-  Pagination,
-  Row,
-  Select,
-  Table,
-} from "antd";
-import React, { useEffect, useState } from "react";
-import { CardContainer } from "../../components/card/CardContainer";
-import Layouts from "../../components/layout/Layout";
+import { Avatar, Badge, Pagination, Row, Select, Table } from "antd";
 import Search from "antd/lib/input/Search";
-import { Option } from "antd/lib/mentions";
+import moment from "moment";
+import { useEffect, useState } from "react";
 import ActionButton from "../../components/button/ActionButton";
-import { EditOutlined } from "@ant-design/icons";
-import color from "../../resource/color";
-import { useLocalStorage } from "../../hook/useLocalStorage";
+import Layouts from "../../components/layout/Layout";
 import { DroneDatasource } from "../../datasource/DroneDatasource";
-import { formatDate } from "../../utilities/TextFormatter";
+import { DronerDroneDatasource } from "../../datasource/DronerDroneDatasource";
+import { UploadImageDatasouce } from "../../datasource/UploadImageDatasource";
 import {
-  DRONER_DRONE_STATUS,
-  DRONER_STATUS,
+  DRONER_DRONE_MAPPING,
   DRONE_STATUS,
   STATUS_COLOR,
 } from "../../definitions/DronerStatus";
-import { DroneBrandListEntity } from "../../entities/DroneBrandEntities";
+import { DroneBrandEntity } from "../../entities/DroneBrandEntities";
+import { DroneEntity } from "../../entities/DroneEntities";
 import { DronerDroneListEntity } from "../../entities/DronerDroneEntities";
-import { DronerDroneDatasource } from "../../datasource/DronerDroneDatasource";
-import moment from "moment";
+import color from "../../resource/color";
+import { formatDate } from "../../utilities/TextFormatter";
+import { EditOutlined, FileTextFilled } from "@ant-design/icons";
+
 function DroneList() {
   const row = 10;
   const [current, setCurrent] = useState(1);
   const [droneList, setDroneList] = useState<DronerDroneListEntity>();
-  const [searchDroneBrand, setSearchDroneBrand] = useState<any>();
-  const [droneBrandId, setDroneBrandId] = useState<any>();
+  const [droneBrand, setDroneBrand] = useState<DroneBrandEntity[]>();
+  const [seriesDrone, setSeriesDrone] = useState<DroneEntity[]>();
+  const [droneBrandId, setDroneBrandId] = useState<string>();
+  const [searchSeriesDrone, setSearchSeriesDrone] = useState<any>();
   const [searchStatus, setSearchStatus] = useState<string>();
   const [searchText, setSearchText] = useState<string>();
 
-  const fetchDroneList = async () => {
+  const fetchDronerDroneList = async () => {
     await DronerDroneDatasource.getDronerDrone(
       current,
       row,
-      searchDroneBrand,
       searchStatus,
+      searchSeriesDrone,
+      droneBrandId,
       searchText
     ).then((res) => {
       setDroneList(res);
     });
   };
+  const fetchDroneList = async () => {
+    await DroneDatasource.getDroneList(current, 500, droneBrandId).then(
+      (res) => {
+        setSeriesDrone(res.data);
+      }
+    );
+  };
   const fetchDroneBrand = async () => {
     await DroneDatasource.getDroneBrandList().then((res) => {
-      setDroneBrandId(res.data);
+      setDroneBrand(res.data);
     });
   };
-
   useEffect(() => {
+    fetchDronerDroneList();
     fetchDroneList();
     fetchDroneBrand();
-  }, [current, searchText, searchStatus, searchDroneBrand]);
+  }, [current, searchStatus, searchSeriesDrone, searchText, droneBrandId]);
 
   const onChangePage = (page: number) => {
     setCurrent(page);
   };
   const changeTextSearch = (value: string) => {
     setSearchText(value);
+    setCurrent(1);
   };
-  const handleDroneBrand = (droneBrandId: string) => {
-    setSearchDroneBrand(droneBrandId);
+  const handleDroneBrand = (droneBrand: string) => {
+    setDroneBrandId(droneBrand);
+    setCurrent(1);
+  };
+  const handleDroneSeries = (seriesDrone: string) => {
+    setSearchSeriesDrone(seriesDrone);
+    setCurrent(1);
   };
   const handleStatus = (status: any) => {
     setSearchStatus(status);
+    setCurrent(1);
   };
   const PageTitle = (
     <>
@@ -79,18 +85,15 @@ function DroneList() {
         className="container d-flex justify-content-between"
         style={{ padding: "10px" }}
       >
-        <div className="col-lg-4">
-          <span style={{ fontSize: 22, fontWeight: "bold" }}>
+        <div className="col-lg-3">
+          <span style={{ fontSize: 20, fontWeight: "bold" }}>
             <span> รายการโดรนเกษตร (Drone List)</span>
           </span>
         </div>
-        <div
-          className="container d-flex justify-content-between"
-          style={{ padding: "8px" }}
-        >
-          <div className="col-lg-5">
+        <div className="container d-flex justify-content-between">
+          <div className="col-lg-4">
             <Search
-              placeholder="ค้นหาเลขตัวถังหรือชื่อนักบินโดรน หรือเบอร์โทร"
+              placeholder="ค้นหาเลขตัวถังหรือชื่อนักบินโดรน"
               className="col-lg-12 p-1"
               onSearch={changeTextSearch}
             />
@@ -112,8 +115,8 @@ function DroneList() {
               allowClear
               onChange={handleDroneBrand}
             >
-              {droneBrandId?.map((item: any) => (
-                <Option value={item.id.toString()}>{item.name}</Option>
+              {droneBrand?.map((item: any) => (
+                <option value={item.id.toString()}>{item.name}</option>
               ))}
             </Select>
           </div>
@@ -132,11 +135,11 @@ function DroneList() {
               }
               placeholder="เลือกรุ่นโดรน"
               allowClear
-              //onChange={handleDroneBrand}
+              onChange={handleDroneSeries}
             >
-              {/* {droneBrandId?.map((item: any) => (
-                <Option value={item.id.toString()}>{item.name}</Option>
-              ))} */}
+              {seriesDrone?.map((item: any) => (
+                <option value={item.id.toString()}>{item.series}</option>
+              ))}
             </Select>
           </div>
           <div className="col">
@@ -159,7 +162,6 @@ function DroneList() {
       title: "วันที่ลงทะเบียน",
       dataIndex: "createdAt",
       key: "createdAt",
-      // width: "20%",
       render: (value: any, row: any, index: number) => {
         return {
           children: (
@@ -178,16 +180,17 @@ function DroneList() {
       title: "ยี่ห้อโดรน",
       dataIndex: "droneBrand",
       key: "droneBrand",
-      // width: "25%",
       render: (value: any, row: any, index: number) => {
+        const filterDrone = row.drone.droneBrandId;
+        let nameDrone = droneBrand?.filter((x) => x.id == filterDrone)[0];
         return {
           children: (
             <div className="container">
-              {/* <span className="text-dark-75  d-block font-size-lg">
-                {row.droneBrand.logoImagePath ? (
+              <span className="text-dark-75  d-block font-size-lg">
+                {nameDrone?.logoImagePath ? (
                   <Avatar
                     size={25}
-                    src={row.droneBrand.logoImagePath}
+                    src={nameDrone.logoImagePath}
                     style={{ marginRight: "5px" }}
                   />
                 ) : (
@@ -195,11 +198,11 @@ function DroneList() {
                     size={25}
                     style={{ color: "#0068F4", backgroundColor: "#EFF2F9" }}
                   >
-                    {row.droneBrand.name.charAt(0)}
+                    {nameDrone?.name.charAt(0)}
                   </Avatar>
                 )}
-               
-              </span> */}
+                {nameDrone?.name}
+              </span>
             </div>
           ),
         };
@@ -209,7 +212,6 @@ function DroneList() {
       title: "รุ่นโดรน",
       dataIndex: "series",
       key: "series",
-      // width: "24%",
       render: (value: any, row: any, index: number) => {
         return {
           children: (
@@ -228,20 +230,18 @@ function DroneList() {
       title: "เลขตัวถัง",
       dataIndex: "serialNo",
       key: "serialNo",
-      // width: "15%",
     },
     {
       title: "ชื่อนักบินโดรน",
       dataIndex: "named",
       key: "named",
-      // width: "15%",
       render: (value: any, row: any, index: number) => {
         return {
           children: (
             <>
               <Row>
                 <span className="text-dark-75  text-hover-primary mb-1 font-size-lg">
-                  {row.droner.firstname + row.droner.lastname}
+                  {row.droner.firstname + " " + row.droner.lastname}
                 </span>
               </Row>
             </>
@@ -251,21 +251,86 @@ function DroneList() {
     },
     {
       title: "ใบอนุญาตนักบิน ",
-      dataIndex: "paper",
-      key: "paper",
-      // width: "15%",
+      dataIndex: "file",
+      key: "file",
+      render: (value: any, row: any, index: number) => {
+        let getLicenseDroner = row.file?.filter(
+          (x: { category: string }) => x.category == "DRONER_LICENSE"
+        );
+        const previewLicenseDroner = async () => {
+          let src = getLicenseDroner[0].path;
+          UploadImageDatasouce.getImage(src.toString()).then((resImg) => {
+            const image = new Image();
+            image.src = resImg.url;
+            const imgWindow = window.open(src);
+            imgWindow?.document.write(image.outerHTML);
+          });
+        };
+        return {
+          children: (
+            <>
+              <Row>
+                <div
+                  className="text-left ps-4"
+                  style={{ cursor: "pointer", color: color.Success, textDecorationLine:"underline"}}
+                >
+                  {getLicenseDroner != false && (
+                    <>
+                      <FileTextFilled />
+                      <text onClick={previewLicenseDroner}>ใบอนุญาตนักบิน</text>
+                    </>
+                  )}
+                </div>
+              </Row>
+            </>
+          ),
+        };
+      },
     },
     {
       title: "ใบอนุญาตโดรน(กสทช.) ",
-      dataIndex: "paperA",
-      key: "paperA",
-      // width: "18%",
+      dataIndex: "licenseDrone",
+      key: "licenseDrone",
+      render: (value: any, row: any, index: number) => {
+        let getLicenseDrone = row.file?.filter(
+          (x: { category: string }) => x.category == "DRONE_LICENSE"
+        );
+        const previewLicenseDrone = async () => {
+          let src = getLicenseDrone[0].path;
+          UploadImageDatasouce.getImage(src.toString()).then((resImg) => {
+            const image = new Image();
+            image.src = resImg.url;
+            const imgWindow = window.open(src);
+            imgWindow?.document.write(image.outerHTML);
+          });
+        };
+        return {
+          children: (
+            <>
+              <Row>
+                <div
+                  className="text-left ps-4"
+                  style={{ cursor: "pointer", color: color.Success,textDecorationLine:"underline" }}
+                >
+                  {getLicenseDrone != false && (
+                    <>
+                      <FileTextFilled />
+                      <text onClick={previewLicenseDrone}>
+                        ใบอนุญาตโดรน(กสทช.)
+                      </text>
+                    </>
+                  )}
+                </div>
+              </Row>
+            </>
+          ),
+        };
+      },
     },
     {
       title: "สถานะ",
       dataIndex: "status",
       key: "status",
-      // width: "13%",
       render: (value: any, row: any, index: number) => {
         const countDay = () => {
           let dateToday: any = moment(Date.now());
@@ -280,7 +345,7 @@ function DroneList() {
             <>
               <span style={{ color: STATUS_COLOR[row.status] }}>
                 <Badge color={STATUS_COLOR[row.status]} />
-                {DRONER_DRONE_STATUS[row.status]}
+                {DRONER_DRONE_MAPPING[row.status]}
                 <br />
               </span>
               <span style={{ color: color.Grey }}>
@@ -295,7 +360,6 @@ function DroneList() {
       title: "",
       dataIndex: "Action",
       key: "Action",
-      // width: "9%",
       render: (value: any, row: any, index: number) => {
         return {
           children: (
@@ -322,15 +386,25 @@ function DroneList() {
         columns={columns}
         dataSource={droneList?.data}
         pagination={false}
-        scroll={{ x: 1300 }}
+        scroll={{ x: 1400 }}
+        rowClassName={(a) =>
+          a.status == "PENDING" &&
+          moment(Date.now()).diff(moment(new Date(a.createdAt)), "day") >= 3
+            ? "PENDING" &&
+              moment(Date.now()).diff(moment(new Date(a.createdAt)), "day") >= 7
+              ? "table-row-older"
+              : "table-row-old"
+            : "table-row-lasted"
+        }
       />
       <div className="d-flex justify-content-between pt-5">
-        <h5>รายการทั้งหมด {droneList?.count} รายการ</h5>
+        <p>รายการทั้งหมด {droneList?.count} รายการ</p>
         <Pagination
           current={current}
           total={droneList?.count}
           onChange={onChangePage}
           pageSize={row}
+          showSizeChanger={false}
         />
       </div>
     </Layouts>